client: fix critical bug
Sometime, when the server send the list of data related to the groups/devices
entities, the client does not receive exactly a block of data.

Example:

The server send

    lana 192.168.48.4/45645	// a block of data related to a device (with the end of line - '\r\n')
    titi 192.168.48.5/45652
    toto 192.168.48.2/45621

But the client can receive

    lana 192.168.48.4/45645
    titi 192.16

This bug occurs because data are put in cache, so in order to prevent this problem
the client read data line by line.

Consequenlty, the block of data is entirely received .

--