graphActivity: logical/displayable object
> Before this commit

GraphActivity used the logical objects → devices (DeviceData list)
in order to display the list of broadcasters and receivers

> Now

GraphActivity generates arrays of strings that will be displayable on
the navigation drawers. Those arrays are stored in the activity in order to
be dealt when the user selects an item related to the device.

--